requirements.txt
slack_sdk
praw
python-dotenv
